rename files in substrate relay
