+++ relays/substrate/src/main.rs
@@ -43,9 +43,9 @@ mod headers_initialize;
 mod messages_lane;
 mod messages_source;
 mod messages_target;
-mod millau_finality_to_rialto;
+mod millau_headers_to_rialto;
 mod millau_messages_to_rialto;
-mod rialto_finality_to_millau;
+mod rialto_headers_to_millau;
 mod rialto_messages_to_millau;
 
 fn main() {
@@ -60,7 +60,7 @@ fn main() {
 async fn run_command(command: cli::Command) -> Result<(), String> {
 	match command {
 		cli::Command::InitBridge(arg) => run_init_bridge(arg).await,
-		cli::Command::RelayHeaders(arg) => run_relay_finality(arg).await,
+		cli::Command::RelayHeaders(arg) => run_relay_headers(arg).await,
 		cli::Command::RelayMessages(arg) => run_relay_messages(arg).await,
 		cli::Command::SendMessage(arg) => run_send_message(arg).await,
 	}
@@ -145,7 +145,7 @@ async fn run_init_bridge(command: cli::InitBridge) -> Result<(), String> {
 	Ok(())
 }
 
-async fn run_relay_finality(command: cli::RelayHeaders) -> Result<(), String> {
+async fn run_relay_headers(command: cli::RelayHeaders) -> Result<(), String> {
 	match command {
 		cli::RelayHeaders::MillauToRialto {
 			millau,
@@ -156,7 +156,7 @@ async fn run_relay_finality(command: cli::RelayHeaders) -> Result<(), String> {
 			let millau_client = millau.into_client().await?;
 			let rialto_client = rialto.into_client().await?;
 			let rialto_sign = rialto_sign.parse()?;
-			millau_finality_to_rialto::run(millau_client, rialto_client, rialto_sign, prometheus_params.into()).await;
+			millau_headers_to_rialto::run(millau_client, rialto_client, rialto_sign, prometheus_params.into()).await;
 		}
 		cli::RelayHeaders::RialtoToMillau {
 			rialto,
@@ -167,7 +167,7 @@ async fn run_relay_finality(command: cli::RelayHeaders) -> Result<(), String> {
 			let rialto_client = rialto.into_client().await?;
 			let millau_client = millau.into_client().await?;
 			let millau_sign = millau_sign.parse()?;
-			rialto_finality_to_millau::run(rialto_client, millau_client, millau_sign, prometheus_params.into()).await;
+			rialto_headers_to_millau::run(rialto_client, millau_client, millau_sign, prometheus_params.into()).await;
 		}
 	}
 	Ok(())
